requirements.txt
Django==1.7
Jinja2==2.7.3
Markdown==2.5
MarkupSafe==0.23
MySQL-python==1.2.3
Pygments==1.6
Sphinx==1.2.3
boto==2.32.1
django-configurations==0.8
django-filter==0.7
django-storages==1.1.8
djangorestframework==2.4.2
docutils==0.12
pytz==2014.7
requests==2.4.1
wsgiref==0.1.2
